requirements.txt
anthropic>=0.38.0
python-dotenv>=1.0.1
Pillow>=10.0.0
PyYAML>=6.0.0
mss>=9.0.0
pyautogui>=0.9.54
pyobjc-core>=10.0
pyobjc-framework-Cocoa>=10.0
pyobjc-framework-Quartz>=10.0
pytest>=7.4.0
